+++ processor/src/connector/spark/loader.py
@@ -1,4 +1,5 @@
 from pyspark.sql import SparkSession
+
 from src.settings.job_settings import JobSettings
 from src.settings.product_settings import ProductSettings