+++ processor/src/connector/spark/standardizer.py
@@ -1,5 +1,6 @@
 from pyspark.sql import SparkSession
-from src.core import processor
+
+from src.connector.spark import processor
 from src.settings.job_settings import JobSettings
 from src.settings.product_settings import ProductSettings