+++ processor/src/settings/job_settings.py
@@ -4,11 +4,12 @@ from typing import Optional
 
 @dataclass
 class JobSettings:
+  connector: Optional[str] = 'spark'
   product_id: Optional[str] = ''
   event_date: Optional[str] = ''
   action: Optional[str] = ''
   models: Optional[str] = ''
-  config_dir: Optional[str] = path.dirname(__file__) + '/../../../../products'
+  config_dir: Optional[str] = path.dirname(__file__) + '/../../../products'
 
   @classmethod
   def parse(cls, args: list[str]):