+++ processor/src/shared/string_utils.py
@@ -1,3 +1,4 @@
+import re
 from typing import Any
 from jinja2 import Environment
 
@@ -7,3 +8,8 @@ def parse(template: str, context: dict[str, Any]):
   except Exception as ex:
     print(f'Cannot parse "{template}" caused by: ${ex}')
     return template
+
+def parse_function_name_and_arguments(func: str) -> tuple[str, list[str]]:
+  match = re.match(r'(.*?)\((.*)\)', func)
+  if match: return (match.group(1).strip(), [match.group(2).strip()])
+  else: return (func, [])